+++ identity/src/main/java/com/stripe/android/identity/ui/ConsentScreen.kt
@@ -45,6 +45,7 @@ import com.stripe.android.identity.networking.models.CollectedDataParam
 import com.stripe.android.identity.networking.models.VerificationPage
 import com.stripe.android.identity.networking.models.VerificationPage.Companion.requireSelfie
 import com.stripe.android.identity.utils.isRemote
+import com.stripe.android.identity.utils.urlWithoutQuery
 import com.stripe.android.identity.viewmodel.IdentityViewModel
 import com.stripe.android.uicore.image.StripeImage
 import com.stripe.android.uicore.image.StripeImageLoader
@@ -158,7 +159,7 @@ private fun SuccessUI(
                         StripeImageLoader(localContext)
                     }
                     StripeImage(
-                        url = merchantLogoUri.toString(),
+                        url = merchantLogoUri.urlWithoutQuery(),
                         imageLoader = imageLoader,
                         contentDescription = stringResource(id = R.string.description_merchant_logo),
                         modifier = Modifier