+++ payments-core/src/main/java/com/stripe/android/Stripe.kt
@@ -4,11 +4,13 @@ import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import androidx.activity.ComponentActivity
+import androidx.annotation.RestrictTo
 import androidx.annotation.Size
 import androidx.annotation.UiThread
 import androidx.annotation.WorkerThread
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
+import com.stripe.android.cards.CardNumber
 import com.stripe.android.core.ApiKeyValidator
 import com.stripe.android.core.ApiVersion
 import com.stripe.android.core.AppInfo
@@ -37,6 +39,7 @@ import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodCreateParams
 import com.stripe.android.model.PersonTokenParams
 import com.stripe.android.model.PiiTokenParams
+import com.stripe.android.model.PossibleBrands
 import com.stripe.android.model.RadarSession
 import com.stripe.android.model.SetupIntent
 import com.stripe.android.model.Source
@@ -1791,6 +1794,48 @@ class Stripe internal constructor(
                     stripeAccount = stripeAccountId
                 )
             )
+        }
+    }
+
+    /**
+     * Retrieve a list of possible brands for the given card number.
+     * Returns an error if the cardNumber length is less than 6 characters.
+     *
+     * @param cardNumber the card number
+     * @param callback a [ApiResultCallback] to receive the result or error
+     */
+    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+    fun retrievePossibleBrands(
+        cardNumber: String,
+        callback: ApiResultCallback<PossibleBrands>
+    ) {
+        CardNumber.Unvalidated(cardNumber).bin ?: run {
+            callback.onError(
+                InvalidRequestException(
+                    message = "cardNumber cannot be less than 6 characters"
+                )
+            )
+            return
+        }
+
+        executeAsync(callback) {
+            val cardMetaData = stripeRepository.retrieveCardMetadata(
+                cardNumber = cardNumber,
+                requestOptions = ApiRequest.Options(
+                    apiKey = publishableKey,
+                    stripeAccount = stripeAccountId
+                )
+            )
+
+            val brands = cardMetaData?.accountRanges?.map {
+                it.brand
+            }
+
+            brands?.let {
+                PossibleBrands(
+                    brands = brands.toSet().toList()
+                )
+            }
         }
     }