+++ payments-core/src/main/java/com/stripe/android/StripeKtx.kt
@@ -1,7 +1,9 @@
 package com.stripe.android
 
 import android.content.Intent
+import androidx.annotation.RestrictTo
 import androidx.annotation.Size
+import com.stripe.android.cards.CardNumber
 import com.stripe.android.core.exception.APIConnectionException
 import com.stripe.android.core.exception.APIException
 import com.stripe.android.core.exception.AuthenticationException
@@ -23,6 +25,7 @@ import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.PaymentMethodCreateParams
 import com.stripe.android.model.PersonTokenParams
 import com.stripe.android.model.PiiTokenParams
+import com.stripe.android.model.PossibleBrands
 import com.stripe.android.model.RadarSession
 import com.stripe.android.model.SetupIntent
 import com.stripe.android.model.Source
@@ -968,4 +971,45 @@ suspend fun Stripe.verifySetupIntentWithMicrodeposits(
             stripeAccount = stripeAccountId
         )
     )
+}
+
+/**
+ * Retrieve a list of possible brands for the given card number.
+ * Returns an error if the cardNumber length is less than 6 characters.
+ *
+ * @param cardNumber the card number
+ *
+ * @return [PossibleBrands] for the given card number
+ */
+@Throws(
+    AuthenticationException::class,
+    InvalidRequestException::class,
+    APIConnectionException::class,
+    APIException::class
+)
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+suspend fun Stripe.retrievePossibleBrands(
+    cardNumber: String
+): PossibleBrands = runApiRequest {
+    CardNumber.Unvalidated(cardNumber).bin ?: throw InvalidRequestException(
+        message = "cardNumber cannot be less than 6 characters"
+    )
+
+    val cardMetaData = stripeRepository.retrieveCardMetadata(
+        cardNumber = cardNumber,
+        requestOptions = ApiRequest.Options(
+            apiKey = publishableKey,
+            stripeAccount = stripeAccountId
+        )
+    )
+
+    val brands = cardMetaData?.accountRanges?.map {
+        it.brand
+    }
+
+    brands?.let {
+        PossibleBrands(
+            brands = brands.toSet().toList()
+        )
+    }
 }