+++ payments-core/src/main/java/com/stripe/android/networking/StripeApiRepository.kt
@@ -9,6 +9,9 @@ import com.stripe.android.FraudDetectionDataRepository
 import com.stripe.android.Stripe
 import com.stripe.android.StripeApiBeta
 import com.stripe.android.cards.Bin
+import com.stripe.android.cards.CardAccountRangeRepository
+import com.stripe.android.cards.CardNumber
+import com.stripe.android.cards.DefaultCardAccountRangeRepositoryFactory
 import com.stripe.android.core.ApiVersion
 import com.stripe.android.core.AppInfo
 import com.stripe.android.core.Logger
@@ -122,6 +125,8 @@ class StripeApiRepository @JvmOverloads internal constructor(
         DefaultAnalyticsRequestExecutor(logger, workContext),
     private val fraudDetectionDataRepository: FraudDetectionDataRepository =
         DefaultFraudDetectionDataRepository(context, workContext),
+    private val cardAccountRangeRepositoryFactory: CardAccountRangeRepository.Factory =
+        DefaultCardAccountRangeRepositoryFactory(context, analyticsRequestExecutor),
     private val paymentAnalyticsRequestFactory: PaymentAnalyticsRequestFactory =
         PaymentAnalyticsRequestFactory(context, publishableKeyProvider, productUsageTokens),
     private val fraudDetectionDataParamsUtils: FraudDetectionDataParamsUtils = FraudDetectionDataParamsUtils(),
@@ -158,8 +163,16 @@ class StripeApiRepository @JvmOverloads internal constructor(
     private val fraudDetectionData: FraudDetectionData?
         get() = fraudDetectionDataRepository.getCached()
 
+    private val cardAccountRangeRepository: CardAccountRangeRepository
+
     init {
         fireFraudDetectionDataRequest()
+
+        cardAccountRangeRepository =
+            cardAccountRangeRepositoryFactory.createWithStripeRepository(
+                stripeRepository = this,
+                publishableKey = publishableKeyProvider()
+            )
 
         CoroutineScope(workContext).launch {
             val httpCacheDir = File(context.cacheDir, "stripe_api_repository_cache")
@@ -1585,6 +1598,24 @@ class StripeApiRepository @JvmOverloads internal constructor(
             params = params,
             options = options,
             analyticsEvent = null,
+        )
+    }
+
+    override suspend fun retrieveCardMetadata(
+        cardNumber: String,
+        requestOptions: ApiRequest.Options
+    ): CardMetadata? {
+        val unvalidatedNumber = CardNumber.Unvalidated(cardNumber)
+
+        val bin = unvalidatedNumber.bin ?: return null
+
+        val accountRanges = cardAccountRangeRepository.getAccountRanges(
+            cardNumber = unvalidatedNumber
+        ) ?: listOf()
+
+        return CardMetadata(
+            bin = bin,
+            accountRanges = accountRanges
         )
     }